// AngularClient/AppSrc/app/shared/Controllers.generated.ts
export const API_BASE_URL = new OpaqueToken('API_BASE_URL');
// ...
export interface IRestaurantClient {
    get(restaurantId: string): Observable<Restaurant | null>;
    getTable(restaurantId: string, tableId: string): Observable<RestaurantTable | null>;
}

@Injectable()
export class RestaurantClient implements IRestaurantClient {
    private http: Http;
    private baseUrl: string;
    protected jsonParseReviver: ((key: string, value: any) => any) | undefined = undefined;

    constructor(@Inject(Http) http: Http, @Optional() @Inject(API_BASE_URL) baseUrl?: string) {
        this.http = http;
        this.baseUrl = baseUrl ? baseUrl : "";
    }

    get(restaurantId: string): Observable<Restaurant | null> {
        let url_ = this.baseUrl + "/api/Restaurant/Get/{restaurantId}";
        if (restaurantId === undefined || restaurantId === null)
            throw new Error("The parameter 'restaurantId' must be defined.");
        url_ = url_.replace("{restaurantId}", encodeURIComponent("" + restaurantId)); 
        url_ = url_.replace(/[?&]$/, "");

        let options_ = {
            method: "get",
            headers: new Headers({
                "Content-Type": "application/json", 
                "Accept": "application/json"
            })
        };

        return this.http.request(url_, options_).flatMap((response_) => {
            return this.processGet(response_);
        }).catch((response_: any) => {
            if (response_ instanceof Response) {
                try {
                    return this.processGet(response_);
                } catch (e) {
                    return <Observable<Restaurant | null>><any>Observable.throw(e);
                }
            } else
                return <Observable<Restaurant | null>><any>Observable.throw(response_);
        });
    }

    protected processGet(response: Response): Observable<Restaurant | null> {
        const status = response.status; 

        let _headers: any = response.headers ? response.headers.toJSON() : {};
        if (status === 200) {
            const _responseText = response.text();
            let result200: any = null;
            let resultData200 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
            result200 = resultData200 ? Restaurant.fromJS(resultData200) : <any>null;
            return Observable.of(result200);
        } else if (status !== 200 && status !== 204) {
            const _responseText = response.text();
            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
        }
        return Observable.of<Restaurant | null>(<any>null);
    }

    getTable(restaurantId: string, tableId: string): Observable<RestaurantTable | null> {
        let url_ = this.baseUrl + "/api/Restaurant/GetTable/{restaurantId}/{tableId}";
        if (restaurantId === undefined || restaurantId === null)
            throw new Error("The parameter 'restaurantId' must be defined.");
        url_ = url_.replace("{restaurantId}", encodeURIComponent("" + restaurantId)); 
        if (tableId === undefined || tableId === null)
            throw new Error("The parameter 'tableId' must be defined.");
        url_ = url_.replace("{tableId}", encodeURIComponent("" + tableId)); 
        url_ = url_.replace(/[?&]$/, "");

        let options_ = {
            method: "get",
            headers: new Headers({
                "Content-Type": "application/json", 
                "Accept": "application/json"
            })
        };

        return this.http.request(url_, options_).flatMap((response_) => {
            return this.processGetTable(response_);
        }).catch((response_: any) => {
            if (response_ instanceof Response) {
                try {
                    return this.processGetTable(response_);
                } catch (e) {
                    return <Observable<RestaurantTable | null>><any>Observable.throw(e);
                }
            } else
                return <Observable<RestaurantTable | null>><any>Observable.throw(response_);
        });
    }

    protected processGetTable(response: Response): Observable<RestaurantTable | null> {
        const status = response.status; 

        let _headers: any = response.headers ? response.headers.toJSON() : {};
        if (status === 200) {
            const _responseText = response.text();
            let result200: any = null;
            let resultData200 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
            result200 = resultData200 ? RestaurantTable.fromJS(resultData200) : <any>null;
            return Observable.of(result200);
        } else if (status !== 200 && status !== 204) {
            const _responseText = response.text();
            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
        }
        return Observable.of<RestaurantTable | null>(<any>null);
    }
}
// ...
export class Restaurant implements IRestaurant {
    name?: string | undefined;
    id: string;

    constructor(data?: IRestaurant) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.name = data["name"];
            this.id = data["id"];
        }
    }

    static fromJS(data: any): Restaurant {
        let result = new Restaurant();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["name"] = this.name;
        data["id"] = this.id;
        return data; 
    }
}

export interface IRestaurant {
    name?: string | undefined;
    id: string;
}

export class Dish implements IDish {
    name?: string | undefined;
    price: number;
    id: string;
    description?: string | undefined;
    imageUrl?: string | undefined;

    constructor(data?: IDish) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.name = data["name"];
            this.price = data["price"];
            this.id = data["id"];
            this.description = data["description"];
            this.imageUrl = data["imageUrl"];
        }
    }

    static fromJS(data: any): Dish {
        let result = new Dish();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["name"] = this.name;
        data["price"] = this.price;
        data["id"] = this.id;
        data["description"] = this.description;
        data["imageUrl"] = this.imageUrl;
        return data; 
    }
}

export interface IDish {
    name?: string | undefined;
    price: number;
    id: string;
    description?: string | undefined;
    imageUrl?: string | undefined;
}

export class RestaurantTable implements IRestaurantTable {
    tablenr: number;
    id: string;

    constructor(data?: IRestaurantTable) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.tablenr = data["tablenr"];
            this.id = data["id"];
        }
    }

    static fromJS(data: any): RestaurantTable {
        let result = new RestaurantTable();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["tablenr"] = this.tablenr;
        data["id"] = this.id;
        return data; 
    }
}

export interface IRestaurantTable {
    tablenr: number;
    id: string;
}

export class Order implements IOrder {
    restaurantId: string;
    orderId: string;
    tableId: string;
    dishId: string;
    price: number;
    orderTime?: Date | undefined;
    servedTime?: Date | undefined;
    orderState: OrderState;

    constructor(data?: IOrder) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.restaurantId = data["restaurantId"];
            this.orderId = data["orderId"];
            this.tableId = data["tableId"];
            this.dishId = data["dishId"];
            this.price = data["price"];
            this.orderTime = data["orderTime"] ? new Date(data["orderTime"].toString()) : <any>undefined;
            this.servedTime = data["servedTime"] ? new Date(data["servedTime"].toString()) : <any>undefined;
            this.orderState = data["orderState"];
        }
    }

    static fromJS(data: any): Order {
        let result = new Order();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["restaurantId"] = this.restaurantId;
        data["orderId"] = this.orderId;
        data["tableId"] = this.tableId;
        data["dishId"] = this.dishId;
        data["price"] = this.price;
        data["orderTime"] = this.orderTime ? this.orderTime.toISOString() : <any>undefined;
        data["servedTime"] = this.servedTime ? this.servedTime.toISOString() : <any>undefined;
        data["orderState"] = this.orderState;
        return data; 
    }
}

export interface IOrder {
    restaurantId: string;
    orderId: string;
    tableId: string;
    dishId: string;
    price: number;
    orderTime?: Date | undefined;
    servedTime?: Date | undefined;
    orderState: OrderState;
}

export enum OrderState {
    None = 0, 
    Ordered = 1, 
    InKitchenQueue = 2, 
    Cooking = 3, 
    Cooked = 4, 
    Served = 5, 
    Payed = 6, 
}